[#3] Move logging package to api
