+++ easy-agent-core/src/main/java/software/fitz/easyagent/core/EasyAgentBootstrap.java
@@ -1,8 +1,8 @@
 package software.fitz.easyagent.core;
 
 import software.fitz.easyagent.api.Plugin;
-import software.fitz.easyagent.core.logging.AgentLogger;
-import software.fitz.easyagent.core.logging.AgentLoggerFactory;
+import software.fitz.easyagent.api.logging.AgentLogger;
+import software.fitz.easyagent.api.logging.AgentLoggerFactory;
 import software.fitz.easyagent.core.transformer.InjectProxyTransformerDelegate;
 import software.fitz.easyagent.core.transformer.TransformerDelegate;
