+++ easy-agent-core/src/main/java/software/fitz/easyagent/core/transformer/InjectProxyTransformerDelegate.java
@@ -5,8 +5,8 @@ import software.fitz.easyagent.core.classloader.AgentClassLoader;
 import software.fitz.easyagent.core.asm.AddProxyClassVisitor;
 import software.fitz.easyagent.api.interceptor.AroundInterceptor;
 import software.fitz.easyagent.core.interceptor.InterceptorRegistry;
-import software.fitz.easyagent.core.logging.AgentLogger;
-import software.fitz.easyagent.core.logging.AgentLoggerFactory;
+import software.fitz.easyagent.api.logging.AgentLogger;
+import software.fitz.easyagent.api.logging.AgentLoggerFactory;
 import software.fitz.easyagent.core.model.InstrumentClass;
 import software.fitz.easyagent.core.model.InterceptorDefinition;
 import org.objectweb.asm.ClassReader;
